cleanup and minor modifications
